api/requirements.txt
Django==2.2.2
django-rest-auth==0.9.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Pillow==6.0.0
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0

# model requirements
absl-py==0.7.1
astor==0.8.0
cycler==0.10.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
imutils==0.5.2
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.1.0
mock==3.0.5
numpy==1.16.3
opencv-python==4.1.0.25
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1
scikit-learn==0.21.1
scipy==1.3.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.4

